requirements.txt
asgiref==3.9.1
cloudinary==1.29.0
dj-database-url==3.0.1
Django==4.2.16
django-allauth==0.51.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.2.0
